Remove technical debt, add recyclerView for routines preview, the user details are now stored after login
Remove redundant code from AESHelper.
Create Routine class which is used to set up the recyclerView.
Create RoutineAdapter for the recyclerView so that it can handle item clicks and deletions. This class creates the view of the recycler in the RoutineFragment.
The user that logged in is now remembered and passed down to the relevant activities and fragments instead of doing database searches based on the email.
Button to add items to the recyclerView from the routines fragment has been added. The button redirects to a new fragment called SurveyFragment which after pressing the send button it will send the input data from to the existing RoutineFragment and a new Routine object will be added to the recycler view list
Rename activity_survey to activity_main
Add survey questions to the string resources