Added new Survey and RoutineView UI and functionality, added routine classes, user and routines can now be passed between activites and fragments, technical debt removed, shared preferences implemented for user, packages added and files moved accordingly
Created packages for repositories, view models, utils and adapters and moved files accordingly.

Moved RoutineFragment recyclerView to use MVVM architecture.
Created Routine classes that are represented in the recyclerView by a title
Added bmi and routine split calculator in the abstract Routine class
Routine abstract class and derived classes: HypertrophyRoutine, FatLossRoutine, StrengthRoutine now implement Parcelable, as a result we can send a Routine Object between activities and fragments  using intents and bundles

Added RoutineViewActivity which is opened when a recyclerView item from RoutineFragment is clicked.
Added intensity and exercise tables and title edit text in the activity_routine_view.xml

Removed email field from database document, we do not need to store duplicates.
User class now implements Parcelable, as a result we can send an User Object between activities and fragments using intent and bundles

Removed RoutineFragment reference from MainActivity because it caused the application to crash when pressing the back button
Added SharedPreferences to remember the user details in MainActivity because when a RoutineViewActivity was started the details were lost due to the activity lifecycle, and when the user pressed the back button to return to the MainActivity the application crashed.

Remove unused DisplayMessageActivity.