Add intensity prescription for fat loss, strength and hypertrophy routine
Remove 1 exercise from the PPL exercise template
Add IntensityTemplateProvider which determines the sets and reps for a training block based on the goal
Add intensity per block 2d array to Routine abstract class and add it to the parcelable object.
Update the exercise search method (calcExercisesForTrainingBlocks) from Routine to look for the exercise's rep range and select it only if the prescribed reps per block is between the said rep range.
Update dataset rep ranges and add exercises.