Removed routine classes that were not needed, added routines to the firebase database, refactored the exercise and routine repository, renamed database classes to have DB at the end
Removed Hypertrophy, Strength and Fat Loss routines since they were not needed.
Split the routine class in RoutineDB and Routine. The RoutineDB is the object that gets pushed to the database, and Routine is the class that's needed to set the exercises and intensity for a routine.

Renamed Exercise class to ExerciseDB due to the Routine class split
Renamed User class to UserDB due to the Routine class split

Added loading dialogs where the was being setup/fetched from the database

Refactored exercise repository to fetch exercises based on types

Added firebase query to the Routine repository which fetches the routines and displays them in the recycler view

Moved the exercise fetch from routine fragment view model to the survey fragment