Added launcher icon, changed application name, fix Routine class to add email, exercisesStr and intensityStr to the parcelable object, Change user visible messages, fixed delete account dialog, add functionality to save file as pdf, fix data inconsistency between different users
Added launcher icon and application logo to login screen

Changed application name to Routine Maker.

Fix Routine class to add email, exercisesStr and intensityStr to the parcelable object. These fields are needed when the routine is updated in the RoutineViewActivity. The update is done by deleting and then adding the updated routine because when the title is changed, the document id has to change as well.

Change user visible messages. Toast messages should not have ! at the end of the sentence.

Fixed delete account functionality by adding a reauthentication dialog. If a firebase user hasn't logged in recently and request the account to be deleted or to update the password, then a reauthentication is required.

Add functionality to save file as pdf. A destination picker tab will be opened where the use can enter the document name using intent, after a destination is picked a picture of the view will pe added to the pdf.

When logging out the routines repository needs to be wiped. This functionality has been added when the user clicks the back button while in Main Activity and when pressing the logout button from the navigation tab.

Add permissions to the android manifest which are requested when the user saves a routine.

Update gradle version