Update gradle, remove technical debt, rework exercise and intensity templates, add bmi and age factors when creating routine, added persistent mechanism, change action bar color, change recycler view style
Update gradle version for playstore

Remove technical debt by removing unused activity and styles xml, removed navigation bar Download and Share buttons as they will not be used there.

Rework exercise and intensity templates to better fit beginners. This was done with the help of a certified personal trainer with years of experience.

Add bmi and age factors when creating routine. BMI now impacts the intensity selection and routine split. Age impacts exercise selection.

Added persistent mechanism to exercises. That means that some exercises can appear in other training blocks if the exercise fits but not in the same block.

Change action bar color from green to the navigation menu header.

Change recycler view style by adding a grey background to the RoutineFragment and add spacing so that cards are not crowded