Reworked the profile fragment, changed the repetition ranges for accessory exercises, moved dialog listener implementation from main activity to profile fragment
Reworked the profile fragment by adding a table with all the current user information. Added a button on the side which brings up a dialog box to input the new values, removed Update Details button as it was no longer required.

Moved dialog listener logic from main activity to profile fragment

Changed the repetition ranges for accessory exercises to 18 because for fat loss routine for people with bmi over 30 crashed the application.