Added text to the routine view activity, changed the activity_routine_view layout, added string resources, generated release apk, added release folder to gitignore
In the RoutineViewAcitivity java I added 4 textviews with dynamic width, the width is set by an async task in the on create method.

Change the intensity table width to wrap content and remove right constraint to keep it to the left of the page.

Remove exercise table right constraint to keep it to the left of the page