Improve user experience, update exercise dataset
Updated incorrect password messages when user tried to change the password or delete account

In routine fragment view model, refactored the delete routine process by removing the adapter and adding a observable Integer which is observed in the RoutineFragment.

After the user completed a survey, the keyboard would remain open. Now after the survey has been completed the keyboard is hidden

Previously, when swapping from the survey fragment to the routine fragment, I created a new routine fragment object and replaced it with that. Now instead of doing that, we pop the fragment back stack instead

In login activity, update the messages when the user enters incorrect details.

In login activity added flag to clear top and finish activity when moving to the main activity. This causes the app to close when the user presses the back button

Changed the on back press functionality from Main activity. Instead of logging out and moving to the login screen, if the user presses the back button in an interval of 2 seconds the app will be closed. This was done to make the user experience better

Added a restraint to the date picker dialog when choosing the date of birth where the user can only be aged between 18 and 100

Added picture in RoutineFragment which shows when the user doesn't have any routines created

Update messages from the routine to make them shorter and direct

Added restraints to first name and last name fields in the register and profile page. you can only add alphabet letters now

Reorder xml in register activity layout to match the layout view